Moved listener classes to util.
