+++ syncany-lib/src/main/java/org/syncany/operations/up/AsyncIndexer.java
@@ -8,6 +8,7 @@ import java.util.Queue;
 import org.syncany.chunk.Deduper;
 import org.syncany.config.Config;
 import org.syncany.database.DatabaseVersion;
+import org.syncany.util.Listener;
 
 /**
  * @author Tim Hegeman
