+++ syncany-util/src/main/java/org/syncany/util/QueueAdderListener.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.syncany.operations.up;
+package org.syncany.util;
 
 import java.util.Queue;
